semana06/dia5/requirements.txt
flask
pymongo
pymongo[srv]
python-dotenv
Flask-PyMongo
mongoengine
flask-mongoengine
marshmallow-mongoengine
six